# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dcba54af1ef4518a5382e85
  title: Wishbox
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /post/:
    post:
      x-skaffolder-id: 5dcbc6a7f1ef4518a5382f9f
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Post
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contents:
                  type: string
                createdAt:
                  type: integer
                isPublic:
                  type: boolean
                title:
                  type: string
                updatedAt:
                  type: integer
    get:
      x-skaffolder-id: 5dcbc6a8f1ef4518a5382fa2
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Post
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /post/{id}:
    delete:
      x-skaffolder-id: 5dcbc6a8f1ef4518a5382fa3
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Post
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcbc6a7f1ef4518a5382f9e
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Post
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    post:
      x-skaffolder-id: 5dcbc6a7f1ef4518a5382f9c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Post
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Post"
  /post/findBycreatedBy/{key}:
    get:
      x-skaffolder-id: 5dcbc6b7f1ef4518a5382fa5
      x-skaffolder-name: findBycreatedBy
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: findBycreatedBy
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBycreatedBy
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBycreatedBy/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBycreatedBy
      tags:
        - Post
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa createdBy da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /post/findBytags/{key}:
    get:
      x-skaffolder-id: 5dcbc6a3f1ef4518a5382f93
      x-skaffolder-name: findBytags
      x-skaffolder-id-resource: 5dcbc494f1ef4518a5382f72
      x-skaffolder-resource: Post
      x-skaffolder-crudAction: findBytags
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBytags
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBytags/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBytags
      tags:
        - Post
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa tags da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    
  #-----  Relation  -----
  /relation/:
    post:
      x-skaffolder-id: 5dcbc23cf1ef4518a5382f59
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-resource: Relation
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Relation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relation"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: integer
    get:
      x-skaffolder-id: 5dcbc23cf1ef4518a5382f60
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-resource: Relation
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Relation
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Relation"
  /relation/{id}:
    delete:
      x-skaffolder-id: 5dcbc23cf1ef4518a5382f5f
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-resource: Relation
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Relation
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dcbc23cf1ef4518a5382f5b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-resource: Relation
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Relation
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relation"
    post:
      x-skaffolder-id: 5dcbc23cf1ef4518a5382f5a
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-resource: Relation
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Relation
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relation"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Relation"
  /relation/findByuser/{key}:
    get:
      x-skaffolder-id: 5dcbc22ef1ef4518a5382f3d
      x-skaffolder-name: findByuser
      x-skaffolder-id-resource: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-resource: Relation
      x-skaffolder-crudAction: findByuser
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findByuser
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findByuser/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findByuser
      tags:
        - Relation
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id della risorsa user da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Relation"
    
  #-----  Tag  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    Post:
      x-skaffolder-id: 5dcbc494f1ef4518a5382f72
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcbc494f1ef4518a5382f6b
      x-skaffolder-url: /post
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        contents:
          type: string
          x-skaffolder-id-attr: 5dcbc494f1ef4518a5382f6d
          x-skaffolder-type: String
          x-skaffolder-required: true
        createdAt:
          type: integer
          x-skaffolder-id-attr: 5dcbc494f1ef4518a5382f6f
          x-skaffolder-type: Date
        isPublic:
          type: boolean
          x-skaffolder-id-attr: 5dcbc494f1ef4518a5382f6e
          x-skaffolder-type: Boolean
          x-skaffolder-required: true
        title:
          type: string
          x-skaffolder-id-attr: 5dcbc494f1ef4518a5382f6c
          x-skaffolder-type: String
          x-skaffolder-required: true
        updatedAt:
          type: integer
          x-skaffolder-id-attr: 5dcbc494f1ef4518a5382f70
          x-skaffolder-type: Date
      x-skaffolder-relations: 
        createdBy:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dcbc494f1ef4518a5382f6b
          x-skaffolder-ent2: 5dcba54af1ef4518a5382e87
        tags:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dcbc494f1ef4518a5382f6b
          x-skaffolder-ent2: 5dcbc504f1ef4518a5382f77
      required:
        - _id
        - contents
        - isPublic
        - title
    Relation:
      x-skaffolder-id: 5dcbc1f5f1ef4518a5382f36
      x-skaffolder-id-db: 5dcba54af1ef4518a5382e86
      x-skaffolder-id-entity: 5dcbc1f5f1ef4518a5382f31
      x-skaffolder-url: /relation
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        status:
          type: integer
          x-skaffolder-id-attr: 5dcbc1f5f1ef4518a5382f32
          x-skaffolder-type: Integer
          x-skaffolder-required: true
          x-skaffolder-enumeration:
            - FOLLOW
            - FOLLOWING
      x-skaffolder-relations: 
        user:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dcbc1f5f1ef4518a5382f31
          x-skaffolder-ent2: 5dcba54af1ef4518a5382e87
      required:
        - _id
        - status
...